GC venn - big refactor
in retrospect I should have split this into two commits..

(1) Implement a rectangle difference function for more straightforward
computation of region areas, and for more straightforward label placement.

(2) Refactor and reorganize bloated RectangleVennUtils into three files: geomtry, layout, and normalizeLayout
